12 tests were covered

+++ bloodtests/models.py
@@ -5,8 +5,8 @@ class Test(models.Model):
     code = models.CharField(max_length=50)
     name = models.CharField(max_length=150)
     unit = models.CharField(max_length=50)
-    upper = models.IntegerField(blank=True, null=True)
-    lower = models.IntegerField(blank=True, null=True)
+    upper = models.FloatField(blank=True, null=True)
+    lower = models.FloatField(blank=True, null=True)
 
 
 class TestResult(models.Model):
